requirements.txt
fastapi==0.109.0
langchain==0.1.0
langchain-groq==0.0.1
python-dotenv==1.0.0
pydantic==2.6.0
httpx==0.27.0
